update to ts-dependency-injection code
1. Directly assign mocks instead of using spyOn()
Use of spyOn() makes the mock more brittle to calls to resetMocks()/clearMocks(), for no real benefit as there's not a use case for undo-ing the mock anyway.

2. Use own implementation of memoize()
Lodash's version pollutes the types such that mocking becomes tedious, we shouldn't have to also use memoize() function when mocking to make the types match!